requirements.txt
aiohttp-cors==0.7.0
aiohttp==3.7.4.post0
aioredis==1.3.1
alembic==1.6.5
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autograd==1.3
backcall==0.2.0
bleach==3.3.0
blessings==1.7
blis==0.7.4
cachetools==4.2.2
catalogue==1.0.0
catboost==0.26
category-encoders==2.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cliff==3.8.0
cloudpickle==1.6.0
cma==2.7
cmaes==0.8.2
cmd2==2.1.2
colorama==0.4.4
colorlog==5.0.1
configspace==0.4.19
cycler==0.10.0
cymem==2.0.5
cython==0.29.23
daal4py==2021.3.0
daal==2021.3.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
deslib==0.3.5
entrypoints==0.3
fastai==2.4
fastcore==1.3.20
fastprogress==1.0.0
filelock==3.0.12
flaml==0.5.6
future==0.18.2
#git+https://github.com/scikit-learn-contrib/boruta_py.git@f2f1e3c2432bb9a96bff4bace3bdb4fdc7e7ce7f#egg=boruta
google-api-core==1.30.0
google-auth==1.32.1
googleapis-common-protos==1.53.0
gpustat==0.6.0
gpy==1.10.0
gpytorch==1.5.0
graphviz==0.16
greenlet==1.1.0
grpcio==1.38.1
hebo==0.1.0
hiredis==2.0.0
hpbandster==0.7.4
idna==2.10
imbalanced-learn==0.8.0
ipykernel==6.0.1
ipython-genutils==0.2.0
ipython==7.25.0
ipywidgets==7.6.3
jedi==0.18.0
jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
liac-arff==2.5.0
lightgbm==3.2.1
llvmlite==0.36.0
lz4==3.1.3
mako==1.1.4
markupsafe==2.0.1
matplotlib-inline==0.1.2
matplotlib==3.4.2
minio==7.1.0
mistune==0.8.4
msgpack==1.0.2
multidict==5.1.0
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.11.0
notebook==6.4.0
numba==0.53.1
numpy==1.21.0
nvidia-ml-py3==7.352.0
opencensus-context==0.1.2
opencensus==0.7.13
openml==0.12.2
optuna==2.8.0
packaging==21.0
pandas==1.2.5
pandocfilters==1.4.3
paramz==0.9.5
parso==0.8.2
patsy==0.5.1
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.3.1
plac==1.1.3
plotly==5.1.0
preshed==3.0.5
prettytable==2.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
py-spy==0.3.7
pyarrow==4.0.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pygments==2.9.0
pymoo==0.4.2.2
pyparsing==2.4.7
pyperclip==1.8.2
pyro4==4.80
pyrsistent==0.18.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
pyyaml==5.4.1
pyzmq==22.1.0
ray[tune]==1.4.1
redis==3.5.3
requests==2.25.1
rsa==4.7.2
scikit-learn-intelex==2021.3.0
scikit-learn==0.24.2
scipy==1.7.0
send2trash==1.7.1
serpent==1.40
shap==0.39.0
six==1.16.0
sklearn-pandas==2.2.0
slicer==0.0.7
spacy==2.3.7
sqlalchemy==1.4.20
srsly==1.0.5
statsmodels==0.12.2
stevedore==3.3.0
tabulate==0.8.9
tbb==2021.3.0
tenacity==8.0.0
tensorboardx==2.4
terminado==0.10.1
testpath==0.5.0
thinc==7.4.5
threadpoolctl==2.1.0
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.36.2
widgetsnbextension==3.5.1
xgboost==1.4.2
xmltodict==0.12.0
xxhash==2.0.2
yarl==1.6.3
